+++ index.html
@@ -167,6 +167,6 @@
     <button id="stats-next">Próximo</button>
   </div>
 
-  <script src="app.js"></script>
+  <script type="module" src="js/main.js"></script>
 </body>
 </html>
